Added Clustering

+++ main.py
@@ -7,6 +7,7 @@ import os
 from model import CVAE
 from train import train_model
 from utils import *
+#from benchmark import *
 
 
 parser = argparse.ArgumentParser('CVAE')
@@ -25,8 +26,9 @@ parser.add_argument('--strides', type=str, default='20')
 parser.add_argument('--padding', type=str, default='')
 
 parser.add_argument('--train', action='store_true')
-parser.add_argument('--load', action='store_true')
-parser.add_argument('--load_path', type=str, default='')
+parser.add_argument('--load', type=str, default='')
+parser.add_argument('--cluster', type=str, default='')
+parser.add_argument('--benchmark', type=str, default='')
 
 parser.add_argument('--save', type=str, default='')
 
@@ -54,9 +56,19 @@ if __name__ == '__main__':
 	
 	##Check for cuda	
 	if cuda_flag:
-		cvae = CVAE(args.latent_size, args.alpha, args.dropout, channels, kernels, strides, padding, args.cuda).cuda().float()
+		if len(args.load) > 0:
+			cvae = torch.load(args.load+'/model')
+			cvae.load_state_dict(torch.load(args.load+'/state_dict'))
+			print('Model loaded')
+		else:
+			cvae = CVAE(args.latent_size, args.alpha, args.dropout, channels, kernels, strides, padding, args.cuda).float().cuda()
 	else:
-		cvae = CVAE(args.latent_size, args.alpha, args.dropout, channels, kernels, strides, padding, args.cuda).float()
+		if len(args.load) > 0:
+			cvae = torch.load(args.load+'/model', map_location='cpu')
+			cvae.load_state_dict(torch.load(args.load+'/state_dict', map_location='cpu'))
+			print('Model loaded')
+		else:
+			cvae = CVAE(args.latent_size, args.alpha, args.dropout, channels, kernels, strides, padding, args.cuda).float()
 	
 	##Train model if in arguments
 	if args.train:
@@ -66,23 +78,42 @@ if __name__ == '__main__':
 	
 	##Save model into directory specified by user
 	if len(args.save) > 0:
-		os.mkdir(os.getcwd()+'/' + str(args.save))
-		torch.save(cvae, './'+str(args.save)+'/modellos')
-		info_file = open('./'+str(args.save)+'/log_file', 'w')
-		info_file.write('### File containing specs for model ###\n')
-		for arg, value in vars(args).items():
-			print(arg, value, sep='\t', file=info_file)
-		print('### MODEL PARAMETERS ###', file=info_file)
-		print(cvae, file=info_file)
+		##If we have not created a directory for the model, then do it
+		if not os.path.isdir(os.getcwd()+'/' + str(args.save)):
+			os.mkdir(os.getcwd()+'/' + str(args.save))
+			info_file = open('./'+str(args.save)+'/log_file', 'w')
+			info_file.write('### File containing specs for model ###\n')
+			for arg, value in vars(args).items():
+				print(arg, value, sep='\t', file=info_file)
+			print('### MODEL PARAMETERS ###', file=info_file)
+			print(cvae, file=info_file)
+		
+		##Check if we have saved a model allready. If we have, then remove it and save the newly trained model
+		if os.path.isfile('./'+str(args.save)+'/model'):
+			os.remove('./'+str(args.save)+'/model')
+			os.remove('./'+str(args.save)+'/state_dict')
+			torch.save(cvae, './'+str(args.save)+'/model')
+			torch.save(cvae.state_dict(), './'+str(args.save)+'/state_dict' )
+		else:
+			torch.save(cvae, './'+str(args.save)+'/model')
+			torch.save(cvae.state_dict(), './'+str(args.save)+'/state_dict' )
+		
+		
+		
 
 		##If model is trained, and a directory to save has been specified then plot the loss.
 		if args.train:
 
 			##Call function from utils.py to PCA plot encoded samples
 			#pca_encode(cvae, dataset, labels, './'+str(args.save)+'/Encode_PCA.pkl')
-			pca_avg_encode(cvae, dataset, labels, args.cuda, './'+str(args.save)+'/Encode_avg_PCA.pkl')
-			with open('./'+str(args.save)+'/loss_plot.pkl','wb') as f:
-				pickle.dump(loss_list, f)
+			if os.path.isfile('./'+str(args.save)+'/Encode_avg_PCA.pkl'):
+				os.remove('./'+str(args.save)+'/Encode_avg_PCA.pkl')
+				pca_avg_encode(cvae, dataset, labels, args.cuda, './'+str(args.save)+'/Encode_avg_PCA.pkl')
+			else:
+				pca_avg_encode(cvae, dataset, labels, args.cuda, './'+str(args.save)+'/Encode_avg_PCA.pkl')
+			if not os.path.isfile('./'+str(args.save)+'/loss_plot.pkl'):
+				with open('./'+str(args.save)+'/loss_plot.pkl','wb') as f:
+					pickle.dump(loss_list, f)
 			'''
 			fig = plt.figure()
 			plt.plot(range(1, len(loss_list[0])+1), loss_list[0], color='r')
@@ -94,7 +125,12 @@ if __name__ == '__main__':
 			plt.ylabel('Loss')
 			fig.savefig('./'+str(args.save)+'/Loss_plot.pdf', bbox_inches='tight')
 			'''
-			
+	if len(args.cluster) > 0:
+		cluster_encoding(cvae, dataset, labels, args.cuda, './'+str(args.cluster)+'/cluster.tsv')
+	
+	if len(args.benchmark) > 0:
+		benchmark(args.benchmark)
+
